security/auth_manager.py
from typing import Dict, Any, Optional
import jwt
import logging
from datetime import datetime, timedelta
from cryptography.fernet import Fernet
import os
from prometheus_client import Counter, Histogram

logger = logging.getLogger(__name__)

# Security metrics
auth_attempts = Counter(
    'auth_attempts_total',
    'Total authentication attempts',
    ['status']
)

token_operations = Counter(
    'token_operations_total',
    'Total token operations',
    ['operation']
)

encryption_time = Histogram(
    'encryption_time_seconds',
    'Time spent on encryption operations',
    ['operation']
)

class AuthManager:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.secret_key = os.getenv('JWT_SECRET')
        self.encryption_key = Fernet(os.getenv('ENCRYPTION_KEY').encode())
        
    async def authenticate(self, credentials: Dict[str, str]) -> Optional[str]:
        """Authenticate user and return JWT token."""
        try:
            # Validate credentials
            if not self._validate_credentials(credentials):
                auth_attempts.labels(status='failed').inc()
                return None
            
            # Generate token
            token = self._generate_token(credentials['username'])
            auth_attempts.labels(status='success').inc()
            token_operations.labels(operation='generate').inc()
            
            return token
            
        except Exception as e:
            logger.error(f"Authentication error: {str(e)}")
            auth_attempts.labels(status='error').inc()
            return None
    
    def _validate_credentials(self, credentials: Dict[str, str]) -> bool:
        """Validate user credentials."""
        # Implement credential validation logic
        return True
    
    def _generate_token(self, username: str) -> str:
        """Generate JWT token."""
        payload = {
            'sub': username,
            'iat': datetime.utcnow(),
            'exp': datetime.utcnow() + timedelta(hours=24)
        }
        return jwt.encode(payload, self.secret_key, algorithm='HS256')
    
    def verify_token(self, token: str) -> Optional[Dict[str, Any]]:
        """Verify JWT token."""
        try:
            token_operations.labels(operation='verify').inc()
            return jwt.decode(token, self.secret_key, algorithms=['HS256'])
        except jwt.InvalidTokenError as e:
            logger.error(f"Token verification failed: {str(e)}")
            return None
    
    def encrypt_data(self, data: str) -> str:
        """Encrypt sensitive data."""
        start_time = datetime.now()
        try:
            encrypted = self.encryption_key.encrypt(data.encode())
            encryption_time.labels(operation='encrypt').observe(
                (datetime.now() - start_time).total_seconds()
            )
            return encrypted.decode()
        except Exception as e:
            logger.error(f"Encryption error: {str(e)}")
            raise
    
    def decrypt_data(self, encrypted_data: str) -> str:
        """Decrypt sensitive data."""
        start_time = datetime.now()
        try:
            decrypted = self.encryption_key.decrypt(encrypted_data.encode())
            encryption_time.labels(operation='decrypt').observe(
                (datetime.now() - start_time).total_seconds()
            )
            return decrypted.decode()
        except Exception as e:
            logger.error(f"Decryption error: {str(e)}")
            raise