requirements.txt
streamlit>=1.28.0
langchain>=0.0.325
langchain_ollama>=0.0.1
selenium>=4.15.2
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
lxml[html_clean]>=4.9.3
html5lib>=1.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.6
asyncio>=3.4.3
prometheus-client>=0.19.0
psutil>=5.9.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
pylint>=3.0.2
mypy>=1.7.0
newspaper3k>=0.2.8
python-json-logger>=2.0.7
redis>=5.0.1
sqlalchemy>=2.0.0
alembic>=1.12.1
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
cryptography>=41.0.5
pydantic>=2.5.1
celery>=5.3.6
flower>=2.0.1